requirements.txt
# Core dependencies
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
pandas>=1.2.0
Pillow>=8.0.0

# CV
opencv-python>=4.5.0

# Other
scikit-learn>=0.24.0
scipy>=1.6.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Progress bars
tqdm>=4.60.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0